+++ CHANGELOG.md
@@ -4,11 +4,18 @@
 
 - When formatting, always run some custom npm-groovy-lint fix rules not corresponding to CodeNarc violations
 - Return CodeNarc and Groovy versions when --version options is called
+- Fixes
+  - Lost indentation when applying some fix rules
 - Updated fix rules:
   - IndentationClosingBraces
   - IndentationComments
+  - SpaceAfterCatch
+  - SpaceAfterIf
 - New fix rules:
   - ClassEndsWithBlankLine
+  - ClassStartsWithNewLine
+  - SpaceAfterFor
+  - SpaceAfterSwitch
 
 ## [3.2.4] 2020-04-03