+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## UNRELEASED
+## [3.3.0] 2020-04-06
 
 - When formatting, always run some custom npm-groovy-lint fix rules not corresponding to CodeNarc violations
 - Return CodeNarc and Groovy versions when --version options is called